+++ core/iterable.ts
