+++ core/rng.ts
