+++ core/shell.ts
